pyproject.toml
[project]
name = "bit_llm"
description = "Binary LLMs"
version = "0.1.0"
dependencies = [
    "torch>=2.2.1",
    "datasets>=2.18.0",
    "wandb>=0.16.3",
    "hydra-core>=1.3.2",
    "lightning>=2.2.0.post0",
]

[tool.setuptools.packages]
find = {}
